+++ tools/Gruntfile.js
@@ -15,75 +15,82 @@ module.exports = function(grunt) {
         instdir + "www/shared/shiny.min.js.map",
         "./temp_concat/shiny.js",
         "./temp_concat/shiny.js.map",
-        instdir + 'www/shared/datepicker/js/bootstrap-datepicker.min.js',
-        instdir + 'www/shared/ionrangeslider/js/ion.rangeSlider.min.js'
+        instdir + "www/shared/datepicker/js/bootstrap-datepicker.min.js",
+        instdir + "www/shared/ionrangeslider/js/ion.rangeSlider.min.js"
       ]
     },
 
     concat: {
       options: {
         process: function(src, filepath) {
-          return '//---------------------------------------------------------------------\n' +
-            '// Source file: ' + filepath + '\n\n' + src;
+          return (
+            "//---------------------------------------------------------------------\n" +
+            "// Source file: " +
+            filepath +
+            "\n\n" +
+            src
+          );
         },
         sourceMap: true
       },
       shiny: {
         src: [
-          js_srcdir + '_start.js',
-          js_srcdir + 'utils.js',
-          js_srcdir + 'browser.js',
-          js_srcdir + 'input_rate.js',
-          js_srcdir + 'shinyapp.js',
-          js_srcdir + 'notifications.js',
-          js_srcdir + 'modal.js',
-          js_srcdir + 'file_processor.js',
-          js_srcdir + 'binding_registry.js',
-          js_srcdir + 'output_binding.js',
-          js_srcdir + 'output_binding_text.js',
-          js_srcdir + 'output_binding_image.js',
-          js_srcdir + 'output_binding_html.js',
-          js_srcdir + 'output_binding_downloadlink.js',
-          js_srcdir + 'output_binding_datatable.js',
-          js_srcdir + 'output_binding_adapter.js',
-          js_srcdir + 'input_binding.js',
-          js_srcdir + 'input_binding_text.js',
-          js_srcdir + 'input_binding_textarea.js',
-          js_srcdir + 'input_binding_password.js',
-          js_srcdir + 'input_binding_number.js',
-          js_srcdir + 'input_binding_checkbox.js',
-          js_srcdir + 'input_binding_slider.js',
-          js_srcdir + 'input_binding_date.js',
-          js_srcdir + 'input_binding_daterange.js',
-          js_srcdir + 'input_binding_select.js',
-          js_srcdir + 'input_binding_radio.js',
-          js_srcdir + 'input_binding_checkboxgroup.js',
-          js_srcdir + 'input_binding_actionbutton.js',
-          js_srcdir + 'input_binding_tabinput.js',
-          js_srcdir + 'input_binding_fileinput.js',
-          js_srcdir + 'init_shiny.js',
-          js_srcdir + 'reactlog.js',
-          js_srcdir + '_end.js'
+          js_srcdir + "_start.js",
+          js_srcdir + "utils.js",
+          js_srcdir + "browser.js",
+          js_srcdir + "input_rate.js",
+          js_srcdir + "shinyapp.js",
+          js_srcdir + "notifications.js",
+          js_srcdir + "modal.js",
+          js_srcdir + "file_processor.js",
+          js_srcdir + "binding_registry.js",
+          js_srcdir + "output_binding.js",
+          js_srcdir + "output_binding_text.js",
+          js_srcdir + "output_binding_image.js",
+          js_srcdir + "output_binding_html.js",
+          js_srcdir + "output_binding_downloadlink.js",
+          js_srcdir + "output_binding_datatable.js",
+          js_srcdir + "output_binding_adapter.js",
+          js_srcdir + "input_binding.js",
+          js_srcdir + "input_binding_text.js",
+          js_srcdir + "input_binding_textarea.js",
+          js_srcdir + "input_binding_password.js",
+          js_srcdir + "input_binding_number.js",
+          js_srcdir + "input_binding_checkbox.js",
+          js_srcdir + "input_binding_slider.js",
+          js_srcdir + "input_binding_date.js",
+          js_srcdir + "input_binding_daterange.js",
+          js_srcdir + "input_binding_select.js",
+          js_srcdir + "input_binding_radio.js",
+          js_srcdir + "input_binding_checkboxgroup.js",
+          js_srcdir + "input_binding_actionbutton.js",
+          js_srcdir + "input_binding_tabinput.js",
+          js_srcdir + "input_binding_fileinput.js",
+          js_srcdir + "init_shiny.js",
+          js_srcdir + "reactlog.js",
+          js_srcdir + "_end.js"
         ],
         // The temp_concat/ directory would have gone under /srcjs/, but the
         // Babel Grunt plugin has trouble finding presets if it operates on a
         // file that's not under the current directory. So we'll put it under
         // ./
-        dest: './temp_concat/shiny.js',
+        dest: "./temp_concat/shiny.js",
         nonull: true
-      },
+      }
     },
 
     "string-replace": {
       version: {
         files: {
-          './temp_concat/shiny.js': './temp_concat/shiny.js'
+          "./temp_concat/shiny.js": "./temp_concat/shiny.js"
         },
         options: {
-          replacements: [{
-            pattern: /{{\s*VERSION\s*}}/g,
-            replacement: pkgInfo().version
-          }]
+          replacements: [
+            {
+              pattern: /{{\s*VERSION\s*}}/g,
+              replacement: pkgInfo().version
+            }
+          ]
         }
       }
     },
@@ -92,80 +99,83 @@ module.exports = function(grunt) {
       options: {
         sourceMap: true,
         compact: false,
-        presets: ['es2015']
+        presets: ["@babel/preset-env"]
       },
       shiny: {
-        src: './temp_concat/shiny.js',
-        dest: instdir + '/www/shared/shiny.js'
+        src: "./temp_concat/shiny.js",
+        dest: instdir + "/www/shared/shiny.js"
       }
     },
 
     eslint: {
       options: {
-        parser: 'babel-eslint',
-        format: require('eslint-stylish-mapped'),
-        extends: 'eslint:recommended',
+        parser: "babel-eslint",
+        format: require("eslint-stylish-mapped"),
+        extends: "eslint:recommended",
         rules: {
           "consistent-return": 1,
           "dot-location": [1, "property"],
-          "eqeqeq": 1,
+          eqeqeq: 1,
           // "no-shadow": 1,
           "no-implicit-globals": 1,
-          "no-restricted-globals": ["error", "name", "length", "top", "location", "parent", "status"],
+          "no-restricted-globals": [
+            "error",
+            "name",
+            "length",
+            "top",
+            "location",
+            "parent",
+            "status"
+          ],
           "no-global-assign": 1,
           "no-undef": 1,
-          "no-unused-vars": [1, {"args": "none"}],
+          "no-unused-vars": [1, { args: "none" }],
           "guard-for-in": 1,
           // "no-use-before-define": [1, {"functions": false}],
-          "semi": [1, "always"]
+          semi: [1, "always"]
         },
-        envs: [
-          "es6",
-          "browser",
-          "jquery"
-        ],
+        envs: ["es6", "browser", "jquery"],
         globals: ["strftime"]
       },
-      shiny: ['./temp_concat/shiny.js']
+      shiny: ["./temp_concat/shiny.js"]
     },
 
     uglify: {
       shiny: {
         options: {
-          banner: '/*! <%= pkg.name %> <%= pkg.version %> | ' +
-                  '(c) 2012-<%= grunt.template.today("yyyy") %> RStudio, Inc. | ' +
-                  'License: <%= pkg.license %> */\n',
+          banner:
+            "/*! <%= pkg.name %> <%= pkg.version %> | " +
+            '(c) 2012-<%= grunt.template.today("yyyy") %> RStudio, Inc. | ' +
+            "License: <%= pkg.license %> */\n",
           sourceMap: true,
           // Base the .min.js sourcemap off of the .js sourcemap created by concat
-          sourceMapIn: instdir + 'www/shared/shiny.js.map',
+          sourceMapIn: instdir + "www/shared/shiny.js.map",
           sourceMapIncludeSources: true
         },
-        src: instdir + 'www/shared/shiny.js',
-        dest: instdir + 'www/shared/shiny.min.js'
+        src: instdir + "www/shared/shiny.js",
+        dest: instdir + "www/shared/shiny.min.js"
       },
       datepicker: {
         src: [
-          instdir + 'www/shared/datepicker/js/bootstrap-datepicker.js',
-          instdir + 'www/shared/datepicker/js/locales/bootstrap-datepicker.*.js'
+          instdir + "www/shared/datepicker/js/bootstrap-datepicker.js",
+          instdir + "www/shared/datepicker/js/locales/bootstrap-datepicker.*.js"
         ],
-        dest: instdir + 'www/shared/datepicker/js/bootstrap-datepicker.min.js'
+        dest: instdir + "www/shared/datepicker/js/bootstrap-datepicker.min.js"
       },
       ionrangeslider: {
-        src: instdir + 'www/shared/ionrangeslider/js/ion.rangeSlider.js',
-        dest: instdir + 'www/shared/ionrangeslider/js/ion.rangeSlider.min.js'
+        src: instdir + "www/shared/ionrangeslider/js/ion.rangeSlider.js",
+        dest: instdir + "www/shared/ionrangeslider/js/ion.rangeSlider.min.js"
       }
     },
 
     watch: {
       shiny: {
-        files: ['<%= concat.shiny.src %>', '../DESCRIPTION'],
-        tasks: [
-          'default'
-        ]
+        files: ["<%= concat.shiny.src %>", "../DESCRIPTION"],
+        tasks: ["default"]
       },
       datepicker: {
-        files: '<%= uglify.datepicker.src %>',
-        tasks: ['newer:uglify:datepicker']
+        files: "<%= uglify.datepicker.src %>",
+        tasks: ["uglify:datepicker"]
       }
     },
 
@@ -174,12 +184,11 @@ module.exports = function(grunt) {
         override: function(detail, include) {
           // If DESCRIPTION is updated, we'll also need to re-minify shiny.js
           // because the min.js file embeds the version number.
-          if (detail.task === 'uglify' && detail.target === 'shiny') {
-            include(isNewer('../DESCRIPTION', detail.time));
+          if (detail.task === "uglify" && detail.target === "shiny") {
+            include(isNewer("../DESCRIPTION", detail.time));
           } else {
             include(false);
           }
-
         }
       }
     }
@@ -215,13 +224,13 @@ module.exports = function(grunt) {
   grunt.initConfig(gruntConfig);
 
   grunt.registerTask('default', [
-    'newer:concat',
-    'newer:string-replace',
+    'concat',
+    'string-replace',
     'validateStringReplace',
-    'newer:eslint',
+    'eslint',
     'configureBabel',
-    'newer:babel',
-    'newer:uglify'
+    'babel',
+    'uglify'
   ]);